runtime: Export SQUASHAPP_NAME and SQUASHAPP_ARGV0
These are useful for applications that need to know the way the
SquashApp was called. Unfortunately when running scripts, it's not
possible to access the original `argv[0]`. When running a script via an
interpreter, `argv[0]` (or `$0`) is always the path of the invoked
script-file.

This fixes the `busybox` example not being able to be symlinked to
alternate names (e.g. `sh`).